Clear todo about AST types testing
Punt on the idea of embedding type info into our serialized AST for
testing---it doesn't buy us much on top of the current setup, and
actually makes writing tests harder as you need to translate from the
tested source to whatever type-serialization conventions we invent.

Instead, go all in on the idea of serializing the AST back to source,
renaming `debugPrint` to `toKsonSource` and documenting accordingly.