Enhance number parsing to include scientific notation
As specified in JSON spec, a decimal number can be followed by 'e' or 'E', then
an optional sign ('-' or '+'), then one or more digits.

When we see a number with this format, it's understood to be in scientific notation.
This is naturally handled by the `toDouble` method already available.

Closes #2.