Improve Location documentation, 1-based clarity
We're still rightfully nervous that Locations wanting to be 1-based is
the root cause of future off-by-one errors, though since they are
created in such a localized manner, crafting a bunch of machinery around
that worry before it's come true seems like overkill.

For now, carefully and (hopefully) clearly document Location to minimize
the risk here.