Add a command line interface
Add the ability to parse KSON on the command line from file or
interactively.  This is a touch bare bones right now, but it positions
us to be able to start building a language server that chats with this
external interface.

Note: this is stubbed out for js and native because moving towards a
“full slice” implementation on the JVM is higher priority right now than
navigating some of the surprising intricacies of a fully mpp app (though
we will get there!)