Move our parser tests to their more natural home
Move the parser tests from `ParserTest` to `KsonTest` since that is
the more natural place to perform these tests.  Put a placeholder
in `ParserTest` noting this for folks who naturally look there
for tests related to the `Parser` class.

In spite of looking like a fairly large change, this is mostly a pure
refactor, with a few simple code sharing changes (creating
`LoggedMessage.print` for instance) to support the refactor.