Begin incorporating error reporting into Parser
Begin handling non-happy path parses and doing building error reporting
by adding a check that ensures a parse has consumed all of the input.

With this change, our existing tests revealed an issue in
`Parser.objectInternals()` where the closing brace was not being
consumed, which is fixed here.

This also results in some `y_` Json suite tests failing because we
have not yet implemented optional commas for objects, so we add these
to `JsonTestSuiteSkipList` for now.