JB Plugin: Lay the foundation for Kson parsing
Create `KsonParserDefinition` and the basic classes needed to properly
satisfy the return types of its interface, and hook that up into
`plugin.xml`.
Note that this includes a near-complete plugin lexer, `KsonLexer`, and
a naive parser definition in `KsonParser`, laying the foundation for
our Jetbrains IDE support.

Testing is still mostly missing at this point, but as we build out and
complete the plugin's lexer and parser, the testing will also be built
out