JB Plugin: beginnings of a parser for Jetbrains
We want to build up to refactoring so that the parsing we do in our
plugin is powered by the underlying official Kson parser, but it'll be
easier to define the right abstractions for that once we have some of
a duplicated plugin parser defined, so that's what we're starting here.

This implements just a tiny bit of the grammar (simple, single
`key: value` kson files), but does exercise nicely the whole parsing
flow.