Parse quotes as part of STRING
This fixes an inconsistency in our grammar (i.e. "STRING" is a literal,
and the grammar does not mention quotes) and our implementation.
Fixing this inconsistency positions us better to refactor our
main parser to be shared with our Jetbrains parser implementation
in `tooling/jetbrains`.

We may yet revisit/refactor this (especially if/when we add `'` as an
alternate string quote char), but this is a clean way to rectify this
for now.