Builder-style parser refactor, Part 1 of 2
Since Kson is a tooling-focused language for humans, we learn from the
tooling masters and port our Parser to the builder/marker
style used in Jetbrains IDEs.

This "Part 1" commit introduces the interfaces needed to implement this
architecture, and validates the abstractions are good by implementing
them in our `tooling/jetbrains` plugin.  "Part 2" will provide the core
implementations.

The main advantages here are the great error-reporting ergonomics
provided by the marker approach (important for our goal to give the user
amazing help when they have invalid kson) and the ability to plug our
official implementation directly into our Jetbrains support.  This also
estimates to be more efficient by avoiding a bunch of mid-parse object
creation.

NOTE: the tests are temporarily in known broken state at this commit
awaiting "Part 2".  This is a worthwhile compromise to make this large
refactor more manageable and reviewable in two sensible chunks.