Get our build playing nice with Apple silicon
Upgrade Kotlin and fix a node compatibility issue to get our project
playing nice one the new Apple cpus.

Note: this also checks in a `yarn.lock` generated by KotlinJS describing
the dependencies used internally by Kotlin.  This is new as of the
Kotlin (`1.6.10`) that we upgrade to in this commit.