Add dash-delimited list syntax
We now support YAML-ish dash-delimited lists in Kson such as:

```
- "list entry"
- "another entry"
- "yet another!"
```

Clean and familiar!  A nice addition to our 💌 to humans maintaining
config files.

Note that this syntax is in addition to the existing bracket-delimited
lists (`[1,2,3]`)

See the grammar description on the `Parser` class for details on the
exact semantics of these lists, but in a nutshell they are simply:
a dash followed by a value is a list entry.

In practice, this should feel very familiar to YAML users, with the one
caveat that since whitespace is _not_ significant in Kson, to embed a
list within a dash-delimited list, you must use a bracket-delimited list:

```
- 42
- ["sub", "list"]
- 3.14
```

Note that no such limitation exists for bracket delimited lists since
their bounds are not ambiguous:

```
[
  ["sub", "list"],
  - "also"
  - "a sublist!"
  - [ "omg subsub list..." ]
]
```