Simplify and improve `Message` handling
Continue to improve how parser-generated `Message`s are handled in the
code, simplifying clients of the Message code and increasing
protections against coding errors in passing the dynamic args.

Some highlights of this refactor:

- we no longer defer formatting, so formatter error stack traces
directly trace to the problem.  If/when doing this work synchronously
in parses is the root cause of a measured performance issue, we can
reintroduce the deferral

- non-nullability of Message args is now enforced

- typo'ed arg names in `doFormat` implementations are now validated
and a helpful error is given if a developer makes a bad lookup

- the dynamic arg handling is now fully encapsulated in `Message.kt`