Fix embed block and embed tag lexing
Embed blocks were not quite lexing completely/correctly for unclosed
blocks.

Move the unclosed block error up out of the lexer and into
the parser, and also improve embed tag lexing to be more permissive,
allowing all characters and whitespace (this is not only more flexible
and user friendly, it is cleaner to implement.  Everybody wins.)