Fix parsing with comments
We were mistakenly including comments in the token stream that we
passed to `KsonBuilder`, and since "COMMENT" tokens are not in the
grammar, parsing would error or fail.

Fix this by eliminating COMMENTS from the token stream unless "gap
free" Lexing is requested, and add a KsonTest to exercise the fix.