Preserve comment metadata in our AST
Build on 201bacd7, adding proper handing of trailing comments and
refactoring to a more holistic approach to preserving comments as
parsed metadata.

As comments can appear pretty much anywhere in a document, and exist
outside the parsed grammar, ensuring the are _all_ captured and
preserved in meaningful way is challenging.  This commit implements
a strategy based on the key insight that a Kson entity must be one
of the following:
* the ROOT element of the document
* a property on an object defined as the ROOT element
* a list element in a list defined as the ROOT element

So, semantically, those are the things that make sense to document, so
when preserving comments into our AST, we associate the comments we find
in the document to the KsonRoot node, an ObjectPropertyNode, or a
ListElementNode they are "around" (preceding, intermingled with, or
trailing on a line).

NOTE that this still neglects one special case: line comments at the
end of a document, as they do are not "around" any entity.  We'll
fill this last gap in a subsequent commit.