Improve number parsing
Move number parsing up out of the lexer into the parser with a dedicated
sub-parser, `NumberParser`.  This parser closely follows the JSON number
grammar making it easy to maintain Kson's promise of being a super-set
of JSON.

Moving the number parsing up out of the parser also makes it easier to
report number parsing errors in tooling/IDEs.