Fix JsonTestSuiteGenerator issues
Trying to upgrade some components of this project had me doing many
build/clean iterations, and I stumbled on flakiness in
`JsonTestSuiteGenerator`:

- the `JsonTestSuiteGenerator.testGenerate` integration test's use of
the actual JSONTestSuite checkout directory was brittle in the
face of a `./gradlew clean` wiping out that directory from
underneath another build process.  Our test coverage for this subsystem
was  greatly improved by bdd8a86025, so we fix this by simply
deleting that test

- one of the ways this brittleness manifested was leaving an empty
directory behind on error, which led to all subsequent builds failing
with a not-at-all helpful error.  We improve the error in this case and
add a test.