Add `buildSrc` build to CI
The ability to build `buildSrc/` as a standalone project has been
quietly broken since upgrading Kotlin to 1.7.21 in 135b2d05d94.  This
went unnoticed because it still builds successfully as part of
root Kson project build, so everything looked fine.

Add a dedicated `./gradlew check` for `buildSrc/` to our CI so this can
never happen again.  Since `buildSrc/` is developed as a standalone
project, it's obviously critical to maintain that build.

NOTE: committing this without fixing the build yet to validate how CI
behaves when this new build step is broken.  Will fix the build in a
subsequent commit before merging into main.