Control character handling for strings
Implement our rules for control characters in strings.  These are based
closely on Json's rule of disallowing all control characters (i.e. they
may only be included in strings in escaped form).  See
https://www.rfc-editor.org/rfc/rfc8259.html#section-7 for more details.

Our rules are similar, except we allow whitespace control characters to
be embedded in our strings.

Note that this also fixed a bug in the Lexer: we had hijacked the null
byte to mean EOF, but that made parsing strings that actually _contain_
the null byte parse incorrectly.