NumberParser cleanup
Inspired by b0c1f050d041, stop hijacking the null byte to mean EOF in
our NumberParser.

Note: because this was private to the scanner in this case and not
actually consulted to check for EOF, there wasn't a true bug in this...
yet.  It's still a bad idea to overload the meaning of the
null byte, so clean this up while we're thinking of it.

Also `getCurrentChar` to `peek` while we're in there.