Formalize handling of empty files
Ensure that trying to parse an empty Kson file produces a
helpful/appropriate error.

Also ensure that the empty file error doesn't bubble up to our editor
where it makes no sense to complain about an empty file while someone
is editing it — it's only a problem once they try to use it.