Integrate JSON-Schema-Test-Suite
In preparation for our work to parse and enforce schemas defined in
[JsonSchema](https://json-schema.org/), extend our build to pull in the
[JSON-Schema-Test-Suite](https://github.com/nst/JSONTestSuite)
project in alongside our checkout of the
[JsonTestSuite](https://github.com/json-schema-org/JSON-Schema-Test-Suite)
project.

This commit also refactors the main git utility, `ensureCleanCheckout`
into its own class.  This isn't strictly necessary, but was done as
part of an early draft of this commit, and I like how it separates
the Json-agnostic Git code into its own class, so I kept it in.

Next up, generating Kson tests from the JSON-Schema-Test-Suite
tests...