Fix quiet object parsing flaw
Fix a flaw in the calculation of the boolean returned by
`objectInternals` where we would return `false` (for "did not parse or
handle anything") in the case where `foundProperties` was `false` but
`allowEmpty` was true.

This bug was "quiet" in the sense that it can't currently be tripped by
the system since the bad return value is only generated by a caller who
does not care about the return value.  But it _was_ strictly wrong and
could become either:

a) confusing to someone trying to understand this code later, OR
b) the root cause of some future bug if/when the callers change

Testing note: no test added for this because it cannot currently
be tripped by the code, and it did not seem worth violating
encapsulation just to test this