Upgrade to JDK 17, improve upgrade experience
Upgrade to JDK 17 while maintaining JDK 11 compile compatibility.

Also improve how/when we error if the build detects a version mismatch.
In particular, previously we sometimes needed to comment out the error
when running `wrapper` to perform the upgrade — now we make sure to
never to hit the version error when `wrapper` is invoked.