Add option to preserve embed tags in Yaml output
When compiling Kson to YAML, embed blocks are normally compiled to
multiline strings since a primary use case for the embed tags is to
facilitate better editor tooling, especially in the case of embedded
code, so to use Kson as an interface on some Yaml that expects code
embedded in multiline strings, Kson's embed blocks must collapse down
to Yaml multiline string.

However, the embed tag metadata is valuable, so this provides an option,
`retainEmbedTags`, to preserve it in our compiled Yaml.  With this
enabled, embed blocks with an embed tag compile to Yaml objects like
this:

```yaml
embedTag: "sql"
embedContent: |
  select from table
```