Fix plugin Kotlin dependency warning, for real
In contrast to d6dcf8259fbf8 (reverted in 8b459c8700424c9), which
quietly broke our plugin, this genuinely fixes the issue.

This was so difficult to address because the warning is actually
buggy for plugins implemented as subprojects, so all my attempts to
address it were futile.  For more context see:
https://github.com/JetBrains/intellij-platform-gradle-plugin/issues/1789.
Also note this comment I made on that issue:
https://github.com/JetBrains/intellij-platform-gradle-plugin/issues/1789#issuecomment-2692475169

So, we fix by explicitly setting `kotlin.stdlib.default.dependency = true`
in the root gradle.properties, which then applies to all subprojects.

This fix is close to ideal, and does no harm. Perhaps it is even
beneficial, since this explicitly states something that I didn't fully
understand until wrestling with all this: the Kotlin stdlib is provided
automatically by Gradle to any project the uses the Kotlin Gradle Plugin,
unless the explicitly opt out by setting
`kotlin.stdlib.default.dependency = false`.  We instead explicitly opt
_in_ with `kotlin.stdlib.default.dependency = true`, which has the
effect of showing the `verifyPluginConfiguration` task that we know
what's going on here and do not need or want a warning.