Indent formatting for Kson
Start building out our desired Kson formatting capabilities by
implementing an indent formatter: this formatter adjusts each line in
the source to have the correct indent for that line's level of
nesting in the document.

Note that this formatter is _only_ an indent formatter.  It does not
adjust any other formatting properties of its input source.  Other
formatting concerns such as line length, list style preferences, etc.
will be implemented on and around the base this indent formatter
provides.

Note that we implement this indent formatter as a scan of the lexed
tokens for the given source, which has a lot of nice properties
including:
- all it requires is a pass of the Lexer, no a full AST construction
- invalid Kson can be indented, which is useful during editing
- the ability to indent invalid source is also a key tool for helping
  users troubleshoot unbalanced braces or surprising nesting: the
  indent-formatted version of a file with unbalanced braces makes
  visible and obvious the true nesting structure of the file
- this can be the basis for determining the next line indent in an
  editor when a user presses enter