Improve Dashed List handling in IndentFormatter
Dashed list elements that are not braced with the optional angle
brackets, `<...>`, need to indent themselves.

So: we add some tracking for who is doing the nested so we can detect
when a dash list element is not in angle brackets, and since dash list
elements do not explicitly end, we need to add `nextUntil` to track
when any nesting added with a list element has been closed so that we
can also close the nest made by the dash list element.

Which is pretty finicky. The nesting counts duping logic from the
`nesting` stack, etc.... Regardless, there's a reasonable approach
here, and this moves our formatter a lot closer to where we want it,
so commiting this milestone and attempting to refactor/improve in
the next commit