Improve IndentFormatter
Previously we had dashed list elements indenting themselves.  This
seemed like the right thing to do because lists written as follows are
very common:

```
my_list:
  - 1
  - 2
  - 3
```

However, I missed the forest for the trees: this "indent a hanging
property value" rule applies to everything, not just dashed lists.
So, for instance, our formatter should also ensure this:

```
my_next_line_string:
  "I'm a next-line string property value that wants to be indented"
```

This commit uses ideas similar to how we were handling list dash
indents to handle these property key/value indents by anchoring a "nest"
on the COLON token when appropriate.

This is more consistent than what we had before and also has the benefit
of fixing the compromise we were living with previously where a
delimiter-free dashed list was _always_ indented, which was okay but not
ideal.