Indent-on-enter for our Jetbrains support
Enhance `KsonEnterHandlerDelegate` to perform next indent-on-enter in
our IDE plugin.  This is delegates down to the Kson formatter itself,
which with a few enhancement is able to expose the new
`getCurrentLineIndentLevel` which implements this indentation
calculation as an only slightly specialized call to the full
`IndentFormatter`, trying to ensure that our auto-indent when a user
enters a new line always agrees with the newline they would get by
invoking a full format.