Kson public interface improvements
Nearly a pure refactor (nearly because now we support comment-free
compiles for Kson and Yaml, see below for details) which improves how
the types click together when making calls to parse Kson.  This should
make things more clear, natural, and discoverable for both developers
working on this code and clients using this code.

Highlights of this change:
- Makes compile configuration more consistent across the layers here
  by moving `CompileTarget` up from `AstNode` into the root `Kson`
  entry point file where it lives more naturally and can be passed
  down to `AstNode`
- Brings consistency to some scattered options like `schemaJson`, and
  `maxNestingLevel` by moving all current compile configuration into
  `CompileTarget` (with the help of the new `CoreCompileConfig`)
- Also makes `preserveComments` configurable, which has the nice
  benefit of not only making the code in AstNode tighter, but also
  means we naturally support a comment-free compile now for both
  Yaml and Kson