Pure refactor: improve Embed Delimiter handling
Replace raw character constants and strings for embed delimiters
(%% and $$) with a sealed class hierarchy to tighten things up.

This gives our embed delimiters a more explicit home which better
encapsulates their relationship and moves us away from the previous
somewhat magic-constant approach.

This was motivated by some potential changes which wanted to define
methods that take an embed delimiter as an argument to perform work
depenedent on which delimiter was used in the original source.
Prior to this change, those methods needed to take `Char` or `String`
args.  Now they may take `EmbedDelim` args, which is much much more
expressive and typesafe.