Fix embed block delimiter handling and escaping
Escaped delimiters in embed blocks are processed into their raw text
when we read the embed content in with `Lexer`, but previous we did
not properly consider escaping when transpiling to our `CompileTargets`,
Kson, Json, and Yaml.  This is particularly important for generating
legal Kson.

Content is now properly escaped in our compiled outputs, and when
re-serializing out to Kson, we delim blocks with whichever delimiter
requires less escaping, and we break ties by using the default
delimiter `%%`

This change also highlighted that `escapeStringLiterals` likely needs
some attention, which we will try to address soon.