Defer embed block unescaping, fix formatting bug
Move embed block content unescaping out of the Lexer, deferring it until
it is needed inside `EmbedBlockNode`.  This moves some cycles off
the very hot path of the Lexer and also fixes an issue with our
formatter:

The formatter re-serializes out the value saved on the lexed token,
so having the unescaped value saved on the token meant we had a bug
where illegal unescaped embed content would be written into the document
upoin formatting.  This changes fixes and tests this issue.