Unified grammar, including optional object braces
An important change towards "finalizing" the Kson language definition
for initial release:

The kson grammar is now unified around a single `ksonValue` value-type
and compositions of that value type into objects and lists.  i.e. a
kson document is a `ksonValue` which is defined as follows:

```
* ksonValue -> plainObject
*           | dashList
*           | delimitedValue # literals plus delimited objects and lists
```

Note that this means we support un-delimited `plainObject`s and
`dashList`s everywhere, and still WITHOUT significant whitespace.
These objects are all structural in the language and our formatter will
fully support deciding what whitespace is appropriate to best
communicate a Kson document's structure.

```
key1: val
key2: val
key3:
  nestedKey: val
```

Some more background/color on this important change:

Previous, we only allowed un-delimited dash-lists, and insisted that
directly nested dash-lists be delimited.  This made the grammar more
complicated, but was safe and mitigated against the one ambiguous case:
where it is impossible to tell which list non-first elements belong to
in a nested list.

```
-
  - "sub list elem1"
  - "sub list elem2??" # Ambiguous which list this belongs to
- "outerListElem??" # Ambiguous which list this belongs to, whitespace
                     # is not significant
```

We now solve this ambiguity by supporting semicolon as a way to denote
the end of an ambiguous un-delimited object or list:

```
-
  - "sub list elem1"
  - "sub list elem2!"; # No ambiguity where this sublist ends
- "outerListElem!" # So this elem is unambiguously in the outer list
```

Then we observe that objects are structurally identical to dash-lists---
the KEYWORD and LIST_DASH play the same role of denoting a `ksonValue`.
So we can extend this same mechanism to support un-delimited objects!
(in fact, we should review the `Parser` implementation to see if there
are areas we can DRY up with this observation so that these can be
maintained and enhanced in sync)

```
key:
  nested_key: val
  "nested key??": val # Ambiguous which object this belongs to
"outer key??": val
```
```
key:
  nested_key: val
  "nested key!": val; # No ambiguity where this object ends
"outer key!": val
```

Design notes: the `;` is only need in cases where there is ambiguity,
so it is not required to close an object or dash list.  This introduces
some risk of "deceptively formatted" objects leading to unintentional
data errors:
```
key:
  nested_key: val
"outer key": val # Oops!  Not actually an outer key
```
We may revisit this decision and make them required. For now though,
since the Kson formatter will be a key tool and encouraged into
everyone's workflows, this risk is hopefully minimal and worthwhile in
exchange for the simpler-looking syntax