Implement language injection
Embed Content is now a valid hosts for language injection!

If the tag of an EmbedBlock is a registered language id, then this
language is automatically injected. If not, it is possible to manually
inject a language.

If the injected language does not contain any escape characters the
injected text and the text in the editor is mapped 1:1. However, when
the injected text contains escape characters (e.g. `$\$` in an
`EMBED_BLOCK` delimited by `$$`) then there is no 1:1 mapping and we need to offset the injected characters.
This offset is handled by the `EmbedContentLiteralTextEscaper` and builds
upon the work done in PR @94.

The inverse operation is done by `EmbedContentElementManipulator` when editing an injected fragment in it's designated editor. In the scenario that we injected the following `EMBED_BLOCK`
```
$$
  injected content
$$
```
And add a $$ to the injected content in the edit fragment
```
  injected content $$
```
This should occur escaped in the original `EMBED_BLOCK`
```
$$
  injected content $\$
$$
```.

The `EmbedContentElementManipulator` generates a `KsonEmbedContent`, using the `KsonElementGenerator`, and replaces the content in the original file.