Implement language injection
Embed Content is now a valid hosts for language injection!

If the tag of an EmbedBlock is a registered language id, then this
language is automatically injected. If not, it is possible to manually
inject a language.

To stay true to the rules of an embed block we only inject in the
trimmed ranges of each line instead of injecting into the whole
`EMBED_CONTENT`. The `KsonIndentHandler` finds these ranges.

The class `EmbedContentLiteralTextEscaper` unescapes delimiters, and
handles offsetting it correctly in the original file.
When there are no escape characters necessary we have a simple 1:1
mapping between the injected text and the text in the editor.
However, when the injected text contains escape characters (e.g. `$\$`
in an `EMBED_BLOCK` delimited by `$$`) then there is no 1:1 mapping and
the extra character needs to be offsetted. This offsetting builds upon
the work done in PR @94.

The inverse operation is done in `updateText`, this is for example
called when auto-completing injected code. To ensure that the completed
line does not break the kson we need to escape the new content and
replacing the injected element. This new `KsonEmbedContent element is
generated using the `KsonElementGenerator`.