Full Kson Formatter
Move away from the token-based formatter approach implemented in
`IndentFormatter` to a full AST-based formatter.

The recent change unifying the Kson grammar around a single `ksonValue`
type that allows brace-free objects and dash lists everywhere demands
too much complexity from the token-based formatter implemented in
`IndentFormatter`, so we move away from that to an AST-based formatter.

The new approach makes implementing precise formatting very easy in
the case of a valid Kson AST---the `toSource` methods simply output
the desired format---but requires extra care when the input Kson is
invalid, so:

We are now able to produce ASTs patched with `[AstNodeError]`
instances to get an AST that represents the structure of invalid Kson
documents, which allows for formatting them.

This commit focuses on implementing the support for the formatting use
cases. A subsequent commit will adjust the formatter output to write
cleaner, less brace-heavy, Kson