pure refactor: split KsonTest into focused files
As the home for effectively all the language-level testing for Kson,
the `KsonTest` class has grown very large.  To give these tests an
organized framework within which to comfortably grow more overtime,
split up the tests along the lines of the grammar (i.e. object tests,
list tests, literal tests, etc.), and then split them further according
to whether they are testing for errors of not.