Fix embed delimiter escape boundary
The design of the two-char (`%%`, `$$`) delimiters for embed blocks
specifically means that a naive exact-match search for the
end delimiter will always succeed.  But our Lexer had some
unnecessary code which Lexed things incorrectly in this case:

```
%%
This embed should end "%\", without a newline at the end.
It should NOT cause an unclosed embed block error %\%%
```

Root cause: this is cruft left over from when we were trying to detect
and handle escapes in the Lexer (mostly refactored away in ede3d3b5).

Simplify away the buggy code and add a regression test.