Introduce the end-dash for dashLists
The unified Kson grammar that was merged in 5923bc11dad732 was a bit
too ambitious in its hopes of only introducing one bit of new `end-`
syntax.  It's obvious now in hindsight, but it's impossible to
unambiguously end mixed plainObjects and dashLists with the same syntax
element because the order in which the mixed elements are closed is
material. They newly added `KsonTest.testParseAmbiguityRegression`
demonstrates a key example in code.

So: we restrict the "end-dot" `.` to being the syntax for explicitly
closing objects, and we introduce the "end-dash" `=` for explicitly
closing lists.  The grammar is otherwise unchanged.

We chose the equals sign for our `end-dash` syntax element due to:
- its visual relation to the list dash: `-` / `=`
- its position on the keyboard: users typing `-` to add list elements
  can easily/ergonomically reach the`=` to close that list