Revamp the language injection
Previously we tried to add the language injection by injecting the whole
`EMBED_CONTENT`, however, that isn't true to the indentation rules in an
embed block.

A workaround was to handle indentation as an empty string "". However,
in that case the highlighter still treated everything between the embed
block delimiters as injected content. Besides that, the logic to
artificially remove the indentation artificially was complex.

The solution lies in first finding the ranges that we ACTUALLY want to
 inject, staying true to the `EMBED_BLOCK` indentation rules.
 And then injecting the language for each of these ranges.

 The `KsonIndentHandler` finds these ranges. This simplifies the code a
 lot since the `KsonLiteralTextEscaper` now only needs to find escape
 delimiters, and we don't need logic to decode the indentation and later
 correctly offset it again.