Fix "Refactor File..." undo bug
Make `KsonExternalFormatter` an `AbstractDocumentFormattingService`
rather than an `AsyncDocumentFormattingService` to work around a bug in
how `AsyncDocumentFormattingService` formatting interacts with the IDE's
undo stack.

Fixes an issue where, for example, if a user:
 - types some chars
 - performs a "Reformat File.."
 - performs an Undo
then the Undo would incorrectly undo one or more of the chars along with
the format.

Note that we weren't using `AsyncDocumentFormattingService` for any
particular reason other than finding it in example when learning
how to implement an `ExternalFormatter` which could delegate down to
`Kson.format`. Turns out `AsyncDocumentFormattingService` is intended
for very slow formatters, which we are not and aspire to never be.
`AbstractDocumentFormattingService` is simpler and more robust.