Unify strings, eliminate "identifier" concept
Early on in Kson's design, I had the idea that unquoted strings would
be semantically meaningful, likely references to named objects, hence
the name "identifier" that was used throughout the code.

This concept is now design cruft, so remove the parts of the code that
still imply it, and simplify things around these being simply unquoted
strings.