Attach Location info to Ast nodes
Weave into our AstNodes (and AstApi) the `Location` of the original
parsed Kson source that was parsed into each node.  This will allow us
to anchor semantic errors detected in further inspections of a parsed
Ast tree back to the source of the error.

The first concrete and critical use case that is coming up is the
planned Json Schema validation work which will be implemented as
analysis on `AstApi` trees, and which will want to report back
line-level schema validation errors to the user.