Language Server Implementation
Implementation of the language server protocol in TypeScript. We provide
the server with capabilities to run Diagnostics, do Formatting and
provide Semantic Tokens for a KsonDocument. Since many IDE's support the
LSP this allows us to more easily extend our tooling and support a
variety of different IDE's.

For parsing and formatting a Kson document we use the Javascript library
that is compiled from our multi-platform Kotlin project.
When a document is created or updated the KsonDocumentsManager parses
the document to an AST. This provides enough information for the
Diagnostic and SemanticToken services to return a report. When the
language server gets a format request we parse the document to Kson.

One note is that in this implementation we always send the full Kson
Document over the wire. This enabled a relatively simple
implementation, but might run into performance issues down the line.
Other implementations I saw send a delta of semantic tokens, or a delta
of the changes.