Make JSON test suite generator output consistent across platforms
Internally, the test suite generator was using `/` as a path separator.
However, when displaying paths it used the platform's path separator
(because that's the default for types like `Path`). As a consequence,
the generated files on Windows were slightly different than on *nix
platforms (which in turn resulted in git thinking the generated files
had been updated).

This change makes sure we use `/` when displaying paths as well, thereby
ensuring that the generated files are exactly the same regardless of the
operating system that runs the `generateJsonTestSuite` command.