Configure Javascript build target
Add a JS build target in `build.gradle.kts` so Kson can be consumed from
Node-based JavaScript and TypeScript projects. TypeScript definitions
are generated automatically for every class annotated with `@JsExport`.

Note 1: this is tested indirectly by the implementation of the
language server protocol. The LSP depdends on a succesful JS compilation
and correct TS bindings.

Note 2: Currently using the tasks jsBrowser and jsNode have the same
output directory (/build/dist/js/(developmentLibrary/productionLibrary).
I have not found an elegant way to solve this, and I'm not sure if it's
an urgent problem, if it's even a problem at all.