Fix: Only mark existing tokens during parsing
Up till this commit it was possible to mark non-existent tokens during
parsing. This could for example occur in an embed block without
EMBED_TAG, or an empty object. This resulted in the situation where the
currentTokenIndex could have a smaller index than the firstTokenIndex.

After commit 3e94dee6, where `Location` is introduced to the `AstApi`,
we could get Runtime errors because the first and last token of a
non-existing token are reversed.

We can fix this by simply disallowing empty markers because they only
really make sense if they are marking something (a real element, or
perhaps a range to mark with an error).

Previously, we parsed empty tokens in:
- invalid list elements
- empty objects
- empty tags
All these have been addressed

Not marking empty tokens also showed a defect in handling empty
embed tags in a `KsonEmbedBlock`. Instead of the possibility to return a
nullable tag we should always return a string.

Co-authored-by: Daniel Marcotte <daniel@marcottemotors.io>