Fix: embed delimiter extraction in KsonEmbedBlock
The assumption in the existing getDelim code was that every
embed block needs to have the token `EMBED_OPEN_DELIM`. However, that is
not true. During parsing we also create an `EMBED_BLOCK_NODE` when the
first delimiter is a `EMBED_OPEN_PARTIAL_DELIM`. What does stay true, is
that an embed delimiter is always the first element in an embed block.

This commit fixes that mistaken assumption by changing the way we
retrieve an embed delimiter.