Only auto-complete if document is free of errors
Our auto-complete should never result in a file with errors. This could
happen on a few occasions, e.g. we would still auto-complete if there
already was a closing angle bracket `>`.
Instead of doing a non-trivial search for a close-delimiter that is not
preceded by a matching open delimiter, the solution I propose is more
straightforward. Auto-completions can only happen when the file is
error-free.
The drawback here is that an error unrelated to the auto-complete
prevent auto-completions too. However, I think this will seldomly happen,
since most users probably fix there errors first, and an auto-complete
that introduces a new error seems more cumbersome and annoying than the
lack of auto-complete.