Change embed block delimiter syntax
Instead of having an opening delimiter and a closing delimiter that both
contain two characters, we decided to make the opening delimiter only
one character.
Currently we implicitly allowed a single opening delimiter
(EMBED_DELIM_PARTIAL), even though it got marked as an error. During
formatting, we would correctly format this to the correct syntax.
This raised the question whether or not it should be marked as an error,
or should be up to the users to choose the syntax.
There are three benefits to use a single-delimiter. First of all, the
code simplifies a bit, we don't need to keep track or distinguish between
EMBED_DELIM_PARTIAL, and EMBED_OPEN_DELIM. Secondly, an auto-complete is
now direct, on the first $ or % we immediately auto-complete the embed
block. Most importantly, we now need less escaping inside an embed block,
now we only need to escape the `EMBED_CLOSE_DELIM`!